+++ app/components/GenerateKeyPopup.tsx
@@ -1,5 +1,6 @@
 import * as React from 'react';
-import { Button, Modal, ModalHeader, ModalBody, ModalFooter, Form, FormGroup, Input, Label } from 'reactstrap';
+import { Button, Modal, ModalHeader, ModalBody, ModalFooter, Form, FormGroup, Input, Label, ButtonGroup } from 'reactstrap';
+import { STRENGTH_OPTIONS } from '../actions/privateKey';
 import FocusingInput from './tools/FocusingInput';
 
 export interface Props {
@@ -7,23 +8,36 @@ export interface Props {
   disabled?: boolean,
   passphrase?: string,
   passphraseRepeat?: string,
+  strength?: number,
   valid?: boolean,
   onChangePassphrase: (value: string) => void,
   onChangePassphraseRepeat: (value: string) => void,
+  onChangeStrength: (value: number) => void,
   onGenerate: () => void,
   onClose: () => void
 }
 
-export default ({ open, disabled, passphrase, passphraseRepeat, valid = true, onChangePassphrase, onChangePassphraseRepeat, onGenerate, onClose }: Props) => (
+export default ({ open, disabled, passphrase, passphraseRepeat, strength, valid = true,
+                  onChangePassphrase, onChangePassphraseRepeat, onChangeStrength, onGenerate, onClose }: Props) => (
   <Modal isOpen={open} toggle={onClose}>
     <ModalHeader toggle={onClose}>Generate private key</ModalHeader>
     <ModalBody>
-      <p>This form allows you to generate a new 2048bit RSA key for use within stash.</p>
+      <p>This form allows you to generate a new RSA key for use within Stash.</p>
       <p>
         While optional, it is <b>highly recommended</b> to protect your key with a <b>passphrase</b>.
         This passphrase will prevent unauthorized access to your secrets, and will be required every time you open Stash.
       </p>
       <Form id="editForm" onSubmit={onGenerate}>
+        <FormGroup>
+          <Label>Strength</Label>
+          <div>
+            <ButtonGroup>
+              {STRENGTH_OPTIONS.map(str =>
+                <Button key={str} outline active={strength === str} onClick={() => onChangeStrength(str)}> {str} bits</Button>
+              )}
+            </ButtonGroup>
+          </div>
+        </FormGroup>
         <FormGroup>
           <Label>Passphrase</Label>
           <FocusingInput type="password" focused valid={valid ? undefined : false} disabled={disabled}